+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This release added form fields (textinput, checkbox, and dropdown), drawing shap
 - Paragraph: Support for paragraph with borders - @ivanlanin GH-294
 - Word2007 Writer : Support for RTL - @Progi1984 GH-331
 - MsDOC Reader: Basic MsDOC Reader - @Progi1984 GH-23 GH-287
+- "absolute" horizontal and vertical positioning of Frame - @basjan GH-302
 
 ### Bugfixes